move NaiveAtlas from GeometryThirdParty to Geometry; don't assume mesh has normals in UVAtlas.Atlas()
(cherry picked from commit abefea6dc8e49c7cfd3d31fb754d1f2fff308ed7)

+++ Geometry/Geometry/Mesh.cs
@@ -1413,7 +1413,8 @@ namespace OPS.Geometry
         /// <summary>
         /// computes target bounds from image pixels
         /// </summary>
-        public void RescaleUVsForTexture(int texWidth, int texHeight, int borderPixels = 2, double growThreshold = 0.1)
+        public void RescaleUVsForTexture(int texWidth, int texHeight, double borderPixels = 2,
+                                         double growThreshold = 0.1)
         {
             var border = borderPixels * Vector2.One;
             var targetMin = Image.PixelToUV(border, texWidth, texHeight);
